+++ packages/updating-element/src/decorators/queryAssignedNodes.ts
@@ -77,10 +77,12 @@ export function queryAssignedNodes(
         const slot = this.renderRoot?.querySelector(slotSelector);
         let nodes = (slot as HTMLSlotElement)?.assignedNodes({flatten});
         if (nodes && selector) {
-          nodes = nodes.filter((node) =>
-            node.nodeType === Node.ELEMENT_NODE && (node as Element).matches
-              ? (node as Element).matches(selector)
-              : legacyMatches.call(node as Element, selector)
+          nodes = nodes.filter(
+            (node) =>
+              node.nodeType === Node.ELEMENT_NODE &&
+              ((node as any).matches
+                ? (node as Element).matches(selector)
+                : legacyMatches.call(node as Element, selector))
           );
         }
         return nodes;